dbsearch: use two-arg websearch_to_tsquery version
We already queried the text search config anyway for GIN and
the version with an explicit config is an immutable function
which ensures the query planner doesn’t do silly things and caches
the evaluation result.
(PostgreSQL 16 seems to execute even the one-arg version only once)

This might help with the repeated reevaluation problem reported in
https://akkoma.dev/AkkomaGang/akkoma/issues/650. Note, another even more
foolproof way to ensure only a single eval happens is using a cross join
with the websearch_to_tsquery result. However this can incur significant
overhead for performing a join operation. Since PostgreSQL 13 [1], the
planner is smart enough to re-inline cross joins with immutable
functions, but before that both to_tsvector variants suffer from this
and as of PostgreSQL 16 the one-arg version still does.

Maybe fixes: https://akkoma.dev/AkkomaGang/akkoma/issues/650