Add banner.png back
It was mistakenly removed as part of the bundled frontend in
82fa766ed7a19167af9e25522a2717910c13f6c7 but actually it
is used as the default value for user profile banners.

Mastodon API typespecs define the relevant field as non-nullable in
https://github.com/mastodon/mastodon/blob/585545d0d5678a6ea4b958af4a3bc6593134d0f6/app/javascript/mastodon/api_types/accounts.ts#L30
and clients do in fact rely on it being non-null, meaning we can't just
omit the value if a user didn't setup a bespoke banner.

Signed-off-by: nicole mikołajczyk <git@mkljczk.pl>